MoneyMentor/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pandas==2.1.3
python-dotenv==1.0.0
pydantic==2.5.2
aiofiles==23.2.1
httpx==0.27.0
groq>=0.9.0
google-genai>=1.0.0
